backend/requirements.txt
fastapi
uvicorn
requests
pdfplumber
pytesseract
Pillow
httpx
python-multipart
